dpsc-atividade-backend/requirements.txt
coverage==7.3.2
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
factory-boy==3.3.0
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
requests==2.32.3